requirements.txt
Django
mysqlclient
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
